# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent hub /v1 contract — single source of truth for hub<->consumer surface",
    "version": "\"2026-09-16\""
  },
  "paths": {
...
    "/v1/model-providers": {
      "get": {
        "description": "the providers the hub manages. One provider is one file in the hub's data dir (`providers/<id>.json`): the endpoint it lives at, the protocol it speaks, the declarations for its models, the selection, and the catalog it last served. A provider is reported as `{id, label, url, api, providerType, providerTypeVersion, providerTypeAvailable, catalogUrl, tokenConfigured, declarations, selection, catalogRevision, createdAt, updatedAt}`; the token itself is never returned. Each provider also reports providerType, providerTypeVersion, providerTypeAvailable. Legacy untyped records are read as custom-compatible v1 without rewriting storage or guessing from endpoint/name.",
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "providers": {
                      "type": "array"
                    },
                    "broken": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "file",
                          "error"
                        ],
                        "properties": {
                          "file": {
                            "type": "string"
                          },
                          "error": {
                            "type": "string"
                          }
                        },
                        "description": "provider files the hub could not load"
                      }
                    }
                  },
                  "required": [
                    "providers",
                    "broken"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "register a provider: `{id?, label?, url, api?, token?, scope?, declarations?}`. The id is generated when absent; a duplicate id is refused (409). `api` is the wire protocol the provider speaks in the harness vocabulary (openai-completions, anthropic-messages, openai-responses, ...); the adapter that writes the harness's own provider entry is what validates it. `declarations` is keyed by model id — what that model accepts (thinking levels, window, limits) — because a provider's model list does not carry any of it. A declaration may also carry `input` — the modalities the model takes (\"text\", \"image\") — because a model that does not state them is not advertised as taking images. A declaration may also carry `name` (the display name; the harness shows it instead of the id) and `cost` ({input, output, cacheRead, cacheWrite} per million tokens, optional tiers). Optional immutable providerType/providerTypeVersion select an installed type. Omitted values select custom-compatible v1 for compatibility. Explicit unknown type/version is rejected before storing credentials. Endpoint requirements and configuration validation belong to the selected type: a type that accepts a caller endpoint requires url; a type that owns its endpoint supplies it and rejects overrides. Fields are the configuration.fields its descriptor declares.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "label": {
                    "type": "string"
                  },
                  "url": {
                    "type": "string"
                  },
                  "api": {
                    "type": "string"
                  },
                  "token": {
                    "type": "string"
                  },
                  "declarations": {
                    "type": "object"
                  },
                  "providerType": {
                    "type": "string"
                  },
                  "providerTypeVersion": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "provider": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "provider"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/model-providers/{id}": {
      "get": {
        "description": "read one provider by id. Each provider also reports providerType, providerTypeVersion, providerTypeAvailable. Legacy untyped records are read as custom-compatible v1 without rewriting storage or guessing from endpoint/name.",
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "provider": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "provider"
                  ]
                }
              }
            }
          }
        }
      },
      "patch": {
        "description": "change a provider: `label`, `scope`, `url`, `api`, `token`, `declarations`. Changing the url, the api or the credential bumps the provider's revision, so the cached catalog reports `stale: true` until it is refreshed.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string"
                  },
                  "url": {
                    "type": "string"
                  },
                  "api": {
                    "type": "string"
                  },
                  "token": {
                    "type": "string"
                  },
                  "declarations": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "provider": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "provider"
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "remove a provider and its stored credential.",
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": true,
                    "id": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "ok",
                    "id"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/model-providers/{id}/models": {
      "get": {
        "description": "the cached catalog without touching the network. Each entry is `{id, name, enabled, available}` plus whatever its declaration states (`thinkingLevels`, `input`, `contextWindow`, `maxTokens`). `stale` is true when the endpoint or credential changed since the fetch; a stale catalog reports every model unavailable.",
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "providerId": {
                      "type": "string"
                    },
                    "fetchedAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "stale": {
                      "type": "boolean"
                    },
                    "models": {
                      "type": "array"
                    }
                  },
                  "required": [
                    "providerId",
                    "fetchedAt",
                    "stale",
                    "models"
                  ]
                }
              }
            }
          }
        }
      },
      "patch": {
        "description": "replace the enabled selection (`{enabledModelIds}`; `[]` disables all, unknown ids are refused). The selection is stored on the provider, not in the fetched catalog, so a refresh never changes it and a model the provider starts serving is selected by default.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "enabledModelIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "enabledModelIds"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "providerId": {
                      "type": "string"
                    },
                    "fetchedAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "stale": {
                      "type": "boolean"
                    },
                    "models": {
                      "type": "array"
                    }
                  },
                  "required": [
                    "providerId",
                    "fetchedAt",
                    "stale",
                    "models"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/harnesses/{id}/extensions": {
      "get": {
        "description": "the extension ids THIS harness may be given (the plugin's shipped set), harness-scoped because a bare id is meaningless across harnesses.",
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "harnessId": {
                      "type": "string"
                    },
                    "available": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "harnessId",
                    "available"
                  ]
                }
              }
            }
          }
        }
      },
      "patch": {
        "description": "select the extensions the hub installs for this harness; an unknown id is refused with the available list.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "extensions": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                },
                "required": [
                  "extensions"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "harnessId": {
                      "type": "string"
                    },
                    "extensions": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "harnessId",
                    "extensions"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
...
}